# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChartSpec": {
        "additionalProperties": false,
        "description": "A bounded, declarative display produced from trusted query results.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 100,
            "title": "Data",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "formatting": {
            "$ref": "#/components/schemas/ChartFormatting"
          },
          "id": {
            "maxLength": 80,
            "title": "Id",
            "type": "string"
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/KPIItem"
            },
            "maxItems": 8,
            "title": "Kpis",
            "type": "array"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ChartSeries"
            },
            "maxItems": 8,
            "title": "Series",
            "type": "array"
          },
          "source_query_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Source Query Ids",
            "type": "array"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "line",
              "bar",
              "stacked_bar",
              "area",
              "pie",
              "scatter",
              "table",
              "kpi"
            ],
            "title": "Type",
            "type": "string"
          },
          "x_field": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Field"
          },
          "y_fields": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Y Fields",
            "type": "array"
          }
        },
        "required": [
          "type",
          "title"
        ],
        "title": "ChartSpec",
        "type": "object"
      },
...
      "CreateRunRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "message": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
      "CreateRunResponse": {
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "const": "running",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "conversation_id",
          "status"
        ],
        "title": "CreateRunResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicRunEventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicRunEvent"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PublicRunEventListResponse",
        "type": "object"
      },
      "PublishReportRequest": {
        "additionalProperties": false,
        "description": "What a caller may choose when publishing. Never what a report says.\n\nRejecting unknown fields rather than ignoring them is the point: a request\ncarrying a figure would otherwise look accepted, and a caller could believe\nit had put a number into a report. Every value in a published document comes\nfrom a query the run executed.",
        "properties": {
          "formats": {
            "items": {
              "enum": [
                "pdf",
                "docx"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Formats",
            "type": "array"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricParameters"
            },
            "maxItems": 8,
            "title": "Metrics",
            "type": "array"
          },
          "narrative": {
            "anyOf": [
              {
                "enum": [
                  "current",
                  "pinned_to_original_period",
                  "excluded_from_refreshed_report"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          },
          "period": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "template": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "template",
          "formats"
        ],
        "title": "PublishReportRequest",
        "type": "object"
      },
      "PublishReportResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/PublishedDocumentResponse"
            },
            "title": "Documents",
            "type": "array"
          },
          "narrative": {
            "default": "current",
            "enum": [
              "current",
              "pinned_to_original_period",
              "excluded_from_refreshed_report"
            ],
            "title": "Narrative",
            "type": "string"
          },
          "rerun_query_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rerun Query Ids",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "template": {
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "template",
          "documents"
        ],
        "title": "PublishReportResponse",
        "type": "object"
      },
...
      "ReportPreview": {
        "additionalProperties": false,
        "description": "Everything a reader needs to judge a report before it is published.",
        "properties": {
          "assignment": {
            "$ref": "#/components/schemas/TemplateAssignment"
          },
          "estimated_page_count": {
            "title": "Estimated Page Count",
            "type": "integer"
          },
          "missing_required_content": {
            "items": {
              "type": "string"
            },
            "title": "Missing Required Content",
            "type": "array"
          },
          "pdf_authoritative_notice": {
            "default": "This preview is generated for review only. The published PDF is the authoritative record of this report's figures and layout. A Word copy is provided for editing and reuse; once edited it is no longer a record of what this run produced.",
            "title": "Pdf Authoritative Notice",
            "type": "string"
          },
          "report": {
            "$ref": "#/components/schemas/CompiledReport"
          },
          "suitability": {
            "$ref": "#/components/schemas/TemplateSuitability"
          },
          "template_name": {
            "title": "Template Name",
            "type": "string"
          },
          "template_title": {
            "title": "Template Title",
            "type": "string"
          }
        },
        "required": [
          "template_name",
          "template_title",
          "report",
          "suitability",
          "assignment",
          "missing_required_content",
          "estimated_page_count"
        ],
        "title": "ReportPreview",
        "type": "object"
      },
      "ReportPreviewRequest": {
        "additionalProperties": false,
        "description": "What a caller may choose to preview. The same shape ``publish`` takes,\nminus the output formats a preview never renders \u2014 never a figure.",
        "properties": {
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricParameters"
            },
            "maxItems": 8,
            "title": "Metrics",
            "type": "array"
          },
          "narrative": {
            "anyOf": [
              {
                "enum": [
                  "current",
                  "pinned_to_original_period",
                  "excluded_from_refreshed_report"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          },
          "period": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "template": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "template"
        ],
        "title": "ReportPreviewRequest",
        "type": "object"
      },
...
      "RunResponse": {
        "properties": {
          "caveats": {
            "items": {
              "type": "string"
            },
            "title": "Caveats",
            "type": "array"
          },
          "charts": {
            "items": {
              "$ref": "#/components/schemas/ChartSpec"
            },
            "title": "Charts",
            "type": "array"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "final_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Response"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunMetricsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AnswerSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "running",
              "completed",
              "failed",
              "waiting_for_approval"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "conversation_id",
          "status",
          "created_at"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "TemplateSuitabilityOverview": {
        "additionalProperties": false,
        "description": "Every template's fit for one run, and which one best fits it.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TemplateSuitability"
            },
            "title": "Items",
            "type": "array"
          },
          "recommended_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Template"
          }
        },
        "required": [
          "items"
        ],
        "title": "TemplateSuitabilityOverview",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Autonomous Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/runs": {
      "post": {
        "operationId": "create_run_api_v1_analytics_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
    "/api/v1/analytics/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_analytics_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
    "/api/v1/analytics/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_events_api_v1_analytics_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "last-event-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
    "/api/v1/analytics/runs/{run_id}/events/history": {
      "get": {
        "description": "Return the same safe public event projection used by SSE for an in-process run.",
        "operationId": "get_event_history_api_v1_analytics_runs__run_id__events_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicRunEventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event History",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
    "/api/v1/analytics/runs/{run_id}/report-preview": {
      "post": {
        "description": "Compile the exact canonical report a publish of the same request would produce.\n\nNo PDF or DOCX is written. Publishing this same request afterward compiles\nthrough the identical path, so the previewed assignment and the published\none always match.",
        "operationId": "preview_report_api_v1_analytics_runs__run_id__report_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Report",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
    "/api/v1/analytics/runs/{run_id}/report-suitability": {
      "get": {
        "description": "Score every template against a run's own displays, and recommend one.\n\nDeterministic assembly only: this reads what the run already produced and\nnever calls a model. A caller is free to publish a different template than\nthe one recommended here, unless that template's own required content is\nstill missing.",
        "operationId": "get_report_suitability_api_v1_analytics_runs__run_id__report_suitability_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSuitabilityOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Suitability",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
    "/api/v1/analytics/runs/{run_id}/reports": {
      "post": {
        "description": "Assemble a completed run into documents. This never calls the model.",
        "operationId": "publish_report_api_v1_analytics_runs__run_id__reports_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Report",
        "tags": [
          "analytics-workbench"
        ]
      }
    },
...
  }
}